backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
pydantic-settings
email-validator
